requirements.txt
asgiref==3.7.2
Django==4.2.5
django-environ==0.11.2
psycopg2==2.9.7
sqlparse==0.4.4
gunicorn==20.1.0
google-cloud-secret-manager==2.12.0